New test

+++ WebAppOppg2/ClientApp/src/app/passengerList/passengerList.component.ts
@@ -6,6 +6,8 @@ import { ActivatedRoute } from '@angular/router';
 import { Passenger } from '../Passenger';
 import { PassengerType } from "../PassengerType";
 import { Ticket } from "../Ticket";
+import { PassengerModal } from './deletePassengerModal';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   templateUrl: "passengerList.component.html"
@@ -17,21 +19,26 @@ export class PassengerList {
   loading: boolean;
   skjema: FormGroup;
   allTickets: Array<Ticket>;
+  passengerForDeleting: string;
+  passengerType: any;
 
   validering = {
-    id: [''],
+    passengerID: [''],
     firstname: [
       null, Validators.compose([Validators.required, Validators.pattern("[a-zA-ZøæåØÆÅ\\-. ]{2,30}")])
     ],
     lastname: [
       null, Validators.compose([Validators.required, Validators.pattern("[a-zA-ZøæåØÆÅ\\-. ]{2,30}")])
     ],
-    email: [''],
-    passengerType: [''],
-    passengerTypeID: ['']
+    email: [
+      null, Validators.compose([Validators.required, Validators.pattern("[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,30}")])
+    ],
+    passengerTypeID: [
+      null, Validators.compose([Validators.required, Validators.pattern("[1-9]{1}")])
+    ]
   }
 
-  constructor(private http: HttpClient, private fb: FormBuilder, private router: Router, private route: ActivatedRoute) {
+  constructor(private http: HttpClient, private fb: FormBuilder, private router: Router, private route: ActivatedRoute, private modalService: NgbModal) {
     this.skjema = fb.group(this.validering);
   }
 
@@ -68,64 +75,74 @@ export class PassengerList {
 
   savePassenger() {
     const savedPas = new Passenger();
-
     savedPas.firstname = this.skjema.value.firstname;
     savedPas.lastname = this.skjema.value.lastname;
     savedPas.email = this.skjema.value.email;
-    savedPas.passengerType = this.skjema.value.passengerTypeID;
-
-    //this.getAllTickets();
+    savedPas.passengerTypeID = Number(this.skjema.value.passengerTypeID);
 
     this.http.post("api/passenger", savedPas)
       .subscribe(retur => {
-        this.allPassengers.push(savedPas)
-        //this.router.navigate(['/passengerList']);
-
-      },
-        error => console.log(error)
-      );
-  };
-
-  deleteOnePassenger(id: number) {
-    this.http.delete("api/passenger/" + id)
-      .subscribe(retur => {
         this.getAllPassengers();
-        this.router.navigate(['/passengerList.component.html']);
+        this.clearFrom();
       },
         error => console.log(error)
       );
   };
 
-  //hvis passasjer slettes så skal alle tickets tilhørende også slettes
+  clearFrom(){
+    this.skjema.reset();
+  }
 
-/*
-  getAllTickets() {
+  delete(id: number){
+    this.getAllTickets(id)
+  }
+
+  getAllTickets(id: number) {
     this.http.get<Ticket[]>("api/order")
       .subscribe(tickets => {
         this.allTickets = tickets;
+        this.checkTickets(id);
         this.loading = false;
       },
         error => console.log(error)
       );
   };
 
-  deleteOneTicket(id: number) {
-    this.http.delete("api/order/" + id)
-      .subscribe(retur => {
-        this.getAllTickets();
-        this.router.navigate(['/admin.component.html']);
-      },
-        error => console.log(error)
-      );
-  };
-
-  deleteTicketsOfPassenger(id: number) {
-    if(id)
-
-
-    for (var ticket of tickets) {
-      this.deleteOneTicket(ticket.id)
+  checkTickets(id: number){
+    let hasTickets = this.allTickets.find(x => x.passengerID === id);
+    if(!hasTickets){
+      this.deleteOnePassenger(id);
+    }else{
+      console.log("passesjer har illett");
     }
-  }*/
+  }
+
+  deleteOnePassenger(id: number) {
+    this.http.get<Passenger>("api/passenger/" + id)
+    .subscribe(passenger => {
+      this.passengerForDeleting = passenger.firstname + " " + passenger.lastname;
+      this.showModalandDelete(id);
+    },
+      error => console.log(error)
+    );
+  }
+
+  showModalandDelete(id: number) {
+    const modalRef = this.modalService.open(PassengerModal);
+    modalRef.componentInstance.name = this.passengerForDeleting;
+    modalRef.result.then(retur => {
+      console.log('Lukket med:' + retur);
+      if (retur == "Slett") {
+        this.http.delete("api/passenger/" + id)
+          .subscribe(retur => {
+            this.getAllPassengers();
+            this.router.navigate(['/passengerList.component.html']);
+          },
+            error => console.log(error)
+          );
+      }
+      this.router.navigate(['/passengerList']);
+    });
+  }
 
 }
